@@ -1,35 +1,51 @@
-import { Component, OnInit,Inject } from '@angular/core';
+import { Component, OnInit, Inject } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 
-import { MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MAT_DIALOG_DATA, MatDialogModule } from '@angular/material/dialog';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatGridListModule } from '@angular/material/grid-list';
+import { MatButtonModule } from '@angular/material/button';
+import { MatTableModule } from '@angular/material/table';
+
 import { Venta } from 'src/app/Interfaces/venta';
 import { DetalleVenta } from 'src/app/Interfaces/detalle-venta';
 
 @Component({
   selector: 'app-modal-detalle-venta',
+  standalone: true,
   templateUrl: './modal-detalle-venta.component.html',
-  styleUrls: ['./modal-detalle-venta.component.css']
+  styleUrls: ['./modal-detalle-venta.component.css'],
+  imports: [
+    CommonModule,
+    FormsModule,
+    MatDialogModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatGridListModule,
+    MatButtonModule,
+    MatTableModule
+  ]
 })
 export class ModalDetalleVentaComponent implements OnInit {
 
-  fechaRegistro:string="";
-  numeroDocumento:string="";
-  tipoPago:string="";
-  total:string="";
-  detalleVenta:DetalleVenta[]=[];
-  columnasTabla:string[]=['producto','cantidad','precio','total']
+  fechaRegistro: string = '';
+  numeroDocumento: string = '';
+  tipoPago: string = '';
+  total: string = '';
+  detalleVenta: DetalleVenta[] = [];
+  columnasTabla: string[] = ['producto', 'cantidad', 'precio', 'total'];
 
   constructor(
-    @Inject(MAT_DIALOG_DATA) public _venta:Venta
-  ) 
-  { 
-    this.fechaRegistro=_venta.fechaRegistro!;
-    this.numeroDocumento=_venta.numeroDocumento!;
-    this.tipoPago=_venta.tipoPago;
-    this.total=_venta.totalTexto;
-    this.detalleVenta=_venta.detalleVenta
+    @Inject(MAT_DIALOG_DATA) public _venta: Venta
+  ) {
+    this.fechaRegistro = _venta.fechaRegistro!;
+    this.numeroDocumento = _venta.numeroDocumento!;
+    this.tipoPago = _venta.tipoPago;
+    this.total = _venta.totalTexto;
+    this.detalleVenta = _venta.detalleVenta;
   }
 
-  ngOnInit(): void {
-  }
-
+  ngOnInit(): void {}
 }
