@@ -1,11 +1,22 @@
-import { Component, OnInit, AfterViewInit,ViewChild } from '@angular/core';
+import { Component, OnInit, AfterViewInit, ViewChild } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormBuilder, FormGroup, Validators, ReactiveFormsModule } from '@angular/forms';
 
-import { FormBuilder,FormGroup,Validators } from '@angular/forms';
+import { MatTableModule } from '@angular/material/table';
+import { MatPaginator, MatPaginatorModule } from '@angular/material/paginator';
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatCardModule } from '@angular/material/card';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatDividerModule } from '@angular/material/divider';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatNativeDateModule, MAT_DATE_FORMATS } from '@angular/material/core';
+import { MatSelectModule } from '@angular/material/select';
+import { MatTooltipModule } from '@angular/material/tooltip';
+
 import { MatTableDataSource } from '@angular/material/table';
-import { MatPaginator } from '@angular/material/paginator';
-import { MatDialog } from '@angular/material/dialog';
-
-import { MAT_DATE_FORMATS } from '@angular/material/core';
 import * as moment from 'moment';
 
 import { ModalDetalleVentaComponent } from '../../Modales/modal-detalle-venta/modal-detalle-venta.component';
@@ -13,84 +24,100 @@ import { Venta } from 'src/app/Interfaces/venta';
 import { VentaService } from 'src/app/Services/venta.service';
 import { UtilidadService } from 'src/app/Reutilizable/utilidad.service';
 
-export const MY_DATA_FORMANTS={
-  parse:{
-    dateInput:'DD/MM/YYYY'
+export const MY_DATA_FORMATS = {
+  parse: {
+    dateInput: 'DD/MM/YYYY',
   },
-  display:{
-    dateInput:'DD/MM/YYYY',
-    monthYearLabel:'MMMM YYYY'
+  display: {
+    dateInput: 'DD/MM/YYYY',
+    monthYearLabel: 'MMMM YYYY',
+    dateA11yLabel: 'LL',
+    monthYearA11yLabel: 'MMMM YYYY'
   }
-}
+};
 
 @Component({
   selector: 'app-historial-venta',
+  standalone: true,
   templateUrl: './historial-venta.component.html',
   styleUrls: ['./historial-venta.component.css'],
-  providers:[
-    {provide:MAT_DATE_FORMATS, useValue:MY_DATA_FORMANTS}
+  providers: [
+    { provide: MAT_DATE_FORMATS, useValue: MY_DATA_FORMATS }
+  ],
+  imports: [
+    CommonModule,
+    ReactiveFormsModule,
+    MatCardModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatTableModule,
+    MatPaginatorModule,
+    MatDialogModule,
+    MatButtonModule,
+    MatIconModule,
+    MatDividerModule,
+    MatDatepickerModule,
+    MatNativeDateModule,
+    MatSelectModule,
+    MatTooltipModule,
+    ModalDetalleVentaComponent
   ]
 })
-export class HistorialVentaComponent implements OnInit,AfterViewInit {
+export class HistorialVentaComponent implements OnInit, AfterViewInit {
 
-  formularioBusqueda:FormGroup;
-  opcionesBusqueda:any[]=[
-    {value:"fecha",descripcion:"Por fechas"},
-    {value:"numero",descripcion:"Numero venta"}
-  ]
-  columnasTabla:string[]=['fechaRegistro','numeroDocumento','tipoPago','total','accion']
-  dataInicio:Venta[]=[]
-  datosListaVenta=new MatTableDataSource(this.dataInicio);
+  formularioBusqueda: FormGroup;
+  opcionesBusqueda: any[] = [
+    { value: "fecha", descripcion: "Por fechas" },
+    { value: "numero", descripcion: "Número venta" }
+  ];
+  columnasTabla: string[] = ['fechaRegistro', 'numeroDocumento', 'tipoPago', 'total', 'accion'];
+  dataInicio: Venta[] = [];
+  datosListaVenta = new MatTableDataSource(this.dataInicio);
   @ViewChild(MatPaginator) paginacionTabla!: MatPaginator;
 
-
   constructor(
-    private fb:FormBuilder,
-    private dialog:MatDialog,
-    private _ventaServicio:VentaService,
-    private _utilidadServicio:UtilidadService
-  ) 
-  { 
-    this.formularioBusqueda=this.fb.group({
-      buscarPor:['fecha'],
-      numero:[''],
-      fechaInicio:[''],
-      fechaFin:[''],
-    })
+    private fb: FormBuilder,
+    private dialog: MatDialog,
+    private _ventaServicio: VentaService,
+    private _utilidadServicio: UtilidadService
+  ) {
+    this.formularioBusqueda = this.fb.group({
+      buscarPor: ['fecha'],
+      numero: [''],
+      fechaInicio: [''],
+      fechaFin: [''],
+    });
 
-    this.formularioBusqueda.get("buscarPor")?.valueChanges.subscribe(value=>{
+    this.formularioBusqueda.get("buscarPor")?.valueChanges.subscribe(value => {
       this.formularioBusqueda.patchValue({
-        numero:"",
-        fechaInicio:"",
-        fechaFin:"",
-      })
-    })
-
-
+        numero: "",
+        fechaInicio: "",
+        fechaFin: "",
+      });
+    });
   }
 
-  ngOnInit(): void {
-  }
+  ngOnInit(): void {}
 
   ngAfterViewInit(): void {
-    this.datosListaVenta.paginator=this.paginacionTabla;
+    this.datosListaVenta.paginator = this.paginacionTabla;
   }
 
-  aplicarFiltroTabla(event: Event){
+  aplicarFiltroTabla(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
-    this.datosListaVenta.filter=filterValue.trim().toLocaleLowerCase();
+    this.datosListaVenta.filter = filterValue.trim().toLocaleLowerCase();
   }
 
-  buscarVentas(){
-    let _fechaInicio:string="";
-    let _fechaFin:string="";
+  buscarVentas() {
+    let _fechaInicio: string = "";
+    let _fechaFin: string = "";
 
-    if(this.formularioBusqueda.value.buscarPor==="fecha"){
-      _fechaInicio=moment(this.formularioBusqueda.value.fechaInicio).format('DD/MM/YYYY');
-      _fechaFin=moment(this.formularioBusqueda.value.fechaFin).format('DD/MM/YYYY');
+    if (this.formularioBusqueda.value.buscarPor === "fecha") {
+      _fechaInicio = moment(this.formularioBusqueda.value.fechaInicio).format('DD/MM/YYYY');
+      _fechaFin = moment(this.formularioBusqueda.value.fechaFin).format('DD/MM/YYYY');
 
-      if(_fechaInicio==="Invalid date" || _fechaFin==="Invalid date"){
-        this._utilidadServicio.mostrarAlerta("Debes ingresar ambas fechas","Oops!");
+      if (_fechaInicio === "Invalid date" || _fechaFin === "Invalid date") {
+        this._utilidadServicio.mostrarAlerta("Debes ingresar ambas fechas", "Oops!");
         return;
       }
     }
@@ -101,23 +128,22 @@ export class HistorialVentaComponent implements OnInit,AfterViewInit {
       _fechaInicio,
       _fechaFin
     ).subscribe({
-      next:(data)=>{
-        if(data.status)
-          this.datosListaVenta=data.value;
+      next: (data) => {
+        if (data.status)
+          this.datosListaVenta.data = data.value;
         else
-          this._utilidadServicio.mostrarAlerta("No se encontraron datos","Oops!");
+          this._utilidadServicio.mostrarAlerta("No se encontraron datos", "Oops!");
       },
-      error:(e)=>{}
-    })
-
+      error: () => {}
+    });
   }
 
-  verDetalleVenta(_venta:Venta){
-    this.dialog.open(ModalDetalleVentaComponent,{
-      data:_venta,
-      disableClose:true,
-      width:'700px'
-    })
+  verDetalleVenta(_venta: Venta) {
+    this.dialog.open(ModalDetalleVentaComponent, {
+      data: _venta,
+      disableClose: true,
+      width: '700px'
+    });
   }
 
 }
