@@ -1,103 +1,123 @@
-import { Component, OnInit, AfterViewInit,ViewChild } from '@angular/core';
+import { Component, OnInit, AfterViewInit, ViewChild } from '@angular/core';
+import { CommonModule } from '@angular/common';
 
-import { MatTableDataSource } from '@angular/material/table';
-import { MatPaginator } from '@angular/material/paginator';
-import { MatDialog } from '@angular/material/dialog';
+import { MatTableModule } from '@angular/material/table';
+import { MatPaginator, MatPaginatorModule } from '@angular/material/paginator';
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatCardModule } from '@angular/material/card';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatDividerModule } from '@angular/material/divider';
 
-import { ModalUsuarioComponent } from '../../Modales/modal-usuario/modal-usuario.component';
 import { Usuario } from 'src/app/Interfaces/usuario';
 import { UsuarioService } from 'src/app/Services/usuario.service';
 import { UtilidadService } from 'src/app/Reutilizable/utilidad.service';
+
+import { MatTableDataSource } from '@angular/material/table';
+import { ModalUsuarioComponent } from '../../Modales/modal-usuario/modal-usuario.component';
+
 import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-usuario',
+  standalone: true,
   templateUrl: './usuario.component.html',
-  styleUrls: ['./usuario.component.css']
+  styleUrls: ['./usuario.component.css'],
+  imports: [
+    CommonModule,
+    MatCardModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatTableModule,
+    MatPaginatorModule,
+    MatDialogModule,
+    MatButtonModule,
+    MatIconModule,
+    MatDividerModule,
+    ModalUsuarioComponent
+  ]
 })
 export class UsuarioComponent implements OnInit, AfterViewInit {
 
-  columnasTabla: string[]=['nombreCompleto','correo','rol','estado','acciones'];
-  dataInicio:Usuario[]=[];
-  dataListaUsuarios= new MatTableDataSource(this.dataInicio);
-  @ViewChild(MatPaginator) paginacionTabla! : MatPaginator;
+  columnasTabla: string[] = ['nombreCompleto', 'correo', 'rol', 'estado', 'acciones'];
+  dataInicio: Usuario[] = [];
+  dataListaUsuarios = new MatTableDataSource(this.dataInicio);
 
+  @ViewChild(MatPaginator) paginacionTabla!: MatPaginator;
 
   constructor(
-    private dialog:MatDialog,
-    private _usuariosServicio:UsuarioService,
-    private _utilidadServicio:UtilidadService
+    private dialog: MatDialog,
+    private _usuariosServicio: UsuarioService,
+    private _utilidadServicio: UtilidadService
   ) { }
-
-  obtenerUsuarios() {
-  this._usuariosServicio.lista().subscribe({
-    next: (data) => {
-      this.dataListaUsuarios.data = data;
-    },
-    error: (e) => {
-      this._utilidadServicio.mostrarAlerta("Error al obtener usuarios", "Error");
-    }
-  });
-}
-
 
   ngOnInit(): void {
     this.obtenerUsuarios();
   }
 
   ngAfterViewInit(): void {
-      this.dataListaUsuarios.paginator=this.paginacionTabla;
+    this.dataListaUsuarios.paginator = this.paginacionTabla;
   }
 
-  aplicarFiltroTabla(event: Event){
+  obtenerUsuarios() {
+    this._usuariosServicio.lista().subscribe({
+      next: (data) => {
+        this.dataListaUsuarios.data = data;
+      },
+      error: () => {
+        this._utilidadServicio.mostrarAlerta("Error al obtener usuarios", "Error");
+      }
+    });
+  }
+
+  aplicarFiltroTabla(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
-    this.dataListaUsuarios.filter=filterValue.trim().toLocaleLowerCase();
+    this.dataListaUsuarios.filter = filterValue.trim().toLocaleLowerCase();
   }
 
-  nuevoUsuario(){
-    this.dialog.open(ModalUsuarioComponent,{
-      disableClose:true
-    }).afterClosed().subscribe(resultado=>{
-      if(resultado===true)this.obtenerUsuarios();
+  nuevoUsuario() {
+    this.dialog.open(ModalUsuarioComponent, {
+      disableClose: true
+    }).afterClosed().subscribe(resultado => {
+      if (resultado === true) this.obtenerUsuarios();
     });
   }
 
-  editarUsuario(usuario:Usuario){
-    this.dialog.open(ModalUsuarioComponent,{
-      disableClose:true,
-      data:usuario
-    }).afterClosed().subscribe(resultado=>{
-      if(resultado===true)this.obtenerUsuarios();
+  editarUsuario(usuario: Usuario) {
+    this.dialog.open(ModalUsuarioComponent, {
+      disableClose: true,
+      data: usuario
+    }).afterClosed().subscribe(resultado => {
+      if (resultado === true) this.obtenerUsuarios();
     });
   }
 
-  eliminarUsuario(usuario:Usuario){
+  eliminarUsuario(usuario: Usuario) {
     Swal.fire({
-      title:'¿Deseas eliminar el usuario?',
+      title: '¿Deseas eliminar el usuario?',
       text: usuario.nombreCompleto,
       icon: 'warning',
-      confirmButtonColor:'#3085d6',
-      confirmButtonText: 'Si, eliminar',
+      confirmButtonColor: '#3085d6',
+      confirmButtonText: 'Sí, eliminar',
       showCancelButton: true,
-      cancelButtonColor:'#d33',
-      cancelButtonText:'No. volver'
-    }).then((resultado)=>{
-      if(resultado.isConfirmed){
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'No, volver'
+    }).then((resultado) => {
+      if (resultado.isConfirmed) {
         this._usuariosServicio.eliminar(usuario.idUsuario).subscribe({
-          next:(data)=>{
-            if(data.status){
-              this._utilidadServicio.mostrarAlerta("El usuario fue eliminado","Listo!");
+          next: (data) => {
+            if (data.status) {
+              this._utilidadServicio.mostrarAlerta("El usuario fue eliminado", "¡Listo!");
               this.obtenerUsuarios();
+            } else {
+              this._utilidadServicio.mostrarAlerta("No se pudo eliminar el usuario", "Error");
             }
-            else
-            this._utilidadServicio.mostrarAlerta("No se pudo eliminar el usuario","Error");
           },
-          error:(e)=>{}
-        })
+          error: () => { }
+        });
       }
-    })
+    });
   }
-
-
-
 }
