@@ -1,50 +1,34 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { LayoutComponent } from './layout.component';
-import { DashBoardComponent } from './Pages/dash-board/dash-board.component';
-import { UsuarioComponent } from './Pages/usuario/usuario.component';
-import { ProductoComponent } from './Pages/producto/producto.component';
-import { VentaComponent } from './Pages/venta/venta.component';
-import { HistorialVentaComponent } from './Pages/historial-venta/historial-venta.component';
-import { ReporteComponent } from './Pages/reporte/reporte.component';
+import { Routes } from '@angular/router';
 
-
-
-const routes: Routes = [
+export const routes: Routes = [
   {
-    path:'',
-    component:LayoutComponent,
-    children:[
-      {
-        path:'dashboard',
-        component:DashBoardComponent
-      },
-      {
-        path:'usuarios',
-        component:UsuarioComponent
-      },
-      {
-        path:'productos',
-        component:ProductoComponent
-      },
-      {
-        path:'venta',
-        component:VentaComponent
-      },
-      {
-        path:'historial_venta',
-        component:HistorialVentaComponent
-      },
-      {
-        path:'reporte',
-        component:ReporteComponent
-      },
-    ]
+    path: '',
+    loadComponent: () =>
+      import('./Pages/dash-board/dash-board.component').then(m => m.DashBoardComponent)
+  },
+  {
+    path: 'usuario',
+    loadComponent: () =>
+      import('./Pages/usuario/usuario.component').then(m => m.UsuarioComponent)
+  },
+  {
+    path: 'producto',
+    loadComponent: () =>
+      import('./Pages/producto/producto.component').then(m => m.ProductoComponent)
+  },
+  {
+    path: 'venta',
+    loadComponent: () =>
+      import('./Pages/venta/venta.component').then(m => m.VentaComponent)
+  },
+  {
+    path: 'historial',
+    loadComponent: () =>
+      import('./Pages/historial-venta/historial-venta.component').then(m => m.HistorialVentaComponent)
+  },
+  {
+    path: 'reporte',
+    loadComponent: () =>
+      import('./Pages/reporte/reporte.component').then(m => m.ReporteComponent)
   }
 ];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class LayoutRoutingModule { }
