@@ -1,11 +1,28 @@
 import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
+import { RouterModule, Router } from '@angular/router';
+import { CommonModule } from '@angular/common';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatIconModule } from '@angular/material/icon';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatListModule } from '@angular/material/list';
+import { MatButtonModule } from '@angular/material/button';
+
 import { Menu } from 'src/app/Interfaces/menu';
 import { MenuService } from 'src/app/Services/menu.service';
 import { UtilidadService } from 'src/app/Reutilizable/utilidad.service';
 
 @Component({
   selector: 'app-layout',
+  standalone: true,
+  imports: [
+    CommonModule,
+    RouterModule, // ✅ Necesario para routerLink y router-outlet
+    MatToolbarModule,
+    MatIconModule,
+    MatSidenavModule,
+    MatListModule,
+    MatButtonModule
+  ],
   templateUrl: './layout.component.html',
   styleUrls: ['./layout.component.css']
 })
@@ -29,13 +46,13 @@ export class LayoutComponent implements OnInit {
       this.rolUsuario = usuario.rol.nombre;
 
       this._menuServicio.lista(usuario.rol.idRol).subscribe({
-  next:(data)=>{
-    this.listaMenus = data;
-  },
-  error:(e)=>{
-    console.error("Error al obtener los menús:", e);
-  }
-});
+        next: (data) => {
+          this.listaMenus = data;
+        },
+        error: (e) => {
+          console.error("Error al obtener los menús:", e);
+        }
+      });
     }
   }
 
